Fixed deleting objects and added collision event generation

+++ engine/objects/PhysicsObject.java
@@ -13,6 +13,7 @@ import java.util.ArrayList;
  * @version 4/27/2020
  */
 public abstract class PhysicsObject extends GameObject {
+    protected String type;
     public Collider collider;
     public Vector2 position;
     public Vector2 velocity;
@@ -24,13 +25,16 @@ public abstract class PhysicsObject extends GameObject {
         velocity = new Vector2(0, 0);
     }
 
+    public String getType() {
+        return type;
+    }
+
     /**
      * @return The position of the object rounded down to the nearest pixel.
      */
     public Vector2 pixelPosition() {
         return position.round();
     }
-
 
     /**
      * Handles collision with solid objects. Moves the {@code PhysicsObject} as far as possible in the desired direction
@@ -58,9 +62,6 @@ public abstract class PhysicsObject extends GameObject {
             Collider.Collision collision = collider.getCollisions();
             collider.setPosition(position);
 
-            // Break if there is no collision at this position
-            if(!collision.collision_found) break;
-
             // Make a list of the solid objects collided with
             ArrayList<PhysicsObject> objects = collision.collided_with;
             ArrayList<Collider> other_colliders = new ArrayList<>();
@@ -69,6 +70,9 @@ public abstract class PhysicsObject extends GameObject {
                     other_colliders.add(o.collider);
                 }
             }
+
+            // Break if there is no collision at this position
+            if(other_colliders.size() == 0) break;
 
             // Get a normal vector from the closest surface of the objects collided with
             Vector2 normal = collider.getNormal(delta_position, other_colliders);
@@ -95,6 +99,8 @@ public abstract class PhysicsObject extends GameObject {
         return normals;
     }
 
+    public abstract void collisionEvent(PhysicsObject object);
+
     @Override
     public void update() {
         double t = Game.stepTimeSeconds();
@@ -105,9 +111,11 @@ public abstract class PhysicsObject extends GameObject {
 
     @Override
     public void delete() {
-        collider.delete();
-        collider = null;
-        super.delete();
+        if(!this.isDeleted()) {
+            super.delete();
+            collider.delete();
+            collider = null;
+        }
     }
 }
